Add app.rb class
